Move packet salting into the user security model
